fix: support temperature events with backfill field
The decoding using ast.literal_eval failed due to a newly introduced field in the temperature events. The boolean isBackfilled could not be decoded by ast because it needs valid python and a all lower case 'false' in json would not work. Changed to using json.loads to unmarshal the json object.